+++ src/index.html
@@ -15,9 +15,6 @@
   </head>
   <body>
     <!-- ❗️❗️❗️ Add partials like so for every page -->
-
-
-    <include src="./partials/all-categories.html"></include>
     <script type="module" src="index.js"></script>
     <header>
       <include src="./partials/header.html"></include>
@@ -27,6 +24,7 @@
         <div class="container">
           <div class="main-menu-wrapper">
             <div id="categories-menu">
+              <include src="./partials/best-sellers.html"></include>
               <include src="./partials/categoriesmenu.html"></include>
             </div>
             <div id="funds">
