+++ src/sass/_best-sellers.scss
@@ -1,0 +1,121 @@
+//  375px phone screen 
+.container {
+  position: relative;
+  max-width: 375px;
+  padding-left: 20px;
+  padding-right: 20px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+// 768px  tablet screen 
+@media screen and (min-width: 768px) {
+  .container {
+    max-width: 768px;
+    padding-left: 32px;
+    padding-right: 32px;
+  }
+}
+// 1440px  desktop screen
+@media screen and (min-width: 1440px) {
+  .container {
+    max-width: 1440px;
+    padding-left: 24px;
+    padding-right: 24px;
+  }
+}
+
+
+
+.title-best-sellers {
+  margin-bottom: 40px;
+  color: #111;
+  text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  font-size: 32px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 38px;
+  letter-spacing: -1.28px;
+}
+
+.title-span {
+  color: #4f2ee8;
+}
+.list {
+}
+.list-names {
+}
+.list-categories {
+  margin-bottom: 40px;
+
+}
+.list-categories-title {
+  color: #b4afaf;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 18px;
+  letter-spacing: 0.42px;
+  text-transform: uppercase;
+  margin-bottom: 18px;
+}
+
+.card-container {
+  display: flex;
+  gap: 24px;
+  flex-wrap: wrap;
+  height: 535px;
+  overflow: hidden;
+}
+
+.item-link {
+  
+}
+
+.card-of-book {
+  position: relative;
+  overflow: hidden;
+  border-radius: 8px;
+  overflow: hidden;
+  margin-bottom: 14px;
+}
+.img-card-title {
+  display: block;
+  width: 335px;
+  height: 485px;
+}
+.book-description {
+  width: 335px;
+}
+.card-title {
+  color: #111;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 18px;
+  letter-spacing: -0.64px;
+  text-transform: uppercase;
+}
+.card-author {
+  color: #b4afaf;
+  font-size: 12px;
+  font-style: italic;
+  font-weight: 400;
+  line-height: 14px; 
+  letter-spacing: -0.48px;
+}
+.book-button {
+  display: block;
+  margin-left: auto;
+  margin-top: 8px;
+  padding: 14px 28px;
+  border: 2px solid #4f2ee8;
+  border-radius: 40px;
+  font-weight: 700;
+  font-size: 14px;
+  line-height: calc(18 / 14);
+  letter-spacing: -0.01em;
+  text-transform: uppercase;
+  cursor: pointer;
+  background-color: transparent;
+}
