+++ src/sass/index.scss
@@ -5,3 +5,4 @@
 @import './modal';
 @import './shopping-list';
 @import './support-ukraine';
+@import './best-sellers';
